Removed Flask folder and separated pdf and text into 2 separate functions
